+++ src/PickerMixin.tsx
@@ -1,10 +1,11 @@
 /* tslint:disable:no-console */
+import React from 'react';
 
 export default {
   select(value) {
-    const children: any = this.toChildrenArray(this.props.children);
+    const children: any = React.Children.toArray(this.props.children);
     for (let i = 0, len = children.length; i < len; i++) {
-      if (this.getChildMember(children[i], 'value') === value) {
+      if (children[i].props.value === value) {
         this.selectByIndex(i);
         return;
       }
@@ -13,7 +14,7 @@ export default {
   },
 
   selectByIndex(index) {
-    if (index < 0 || index >= this.toChildrenArray(this.props.children).length || !this.itemHeight) {
+    if (index < 0 || index >= React.Children.count(this.props.children) || !this.itemHeight) {
       return;
     }
     this.scrollTo(index * this.itemHeight);
@@ -27,11 +28,11 @@ export default {
     } else {
       index = floor;
     }
-    const children = this.toChildrenArray(this.props.children);
+    const children = React.Children.toArray(this.props.children);
     index = Math.min(index, children.length - 1);
     const child: any = children[index];
     if (child) {
-      this.fireValueChange(this.getChildMember(child, 'value'));
+      this.fireValueChange(child.props.value);
     } else if (console.warn) {
       console.warn('child not found', children, index);
     }