+++ src/isChildrenEqual.tsx
@@ -1,20 +1,17 @@
-export function isEmptyArray(a) {
-  return !a || !a.length;
-}
+import React from 'react';
 
-export default function isChildrenEqual(c1, c2, pure) {
-  if (isEmptyArray(c1) && isEmptyArray(c2)) {
+export default function isChildrenEqual(_c1, _c2) {
+  if (_c1 === _c1) {
     return true;
   }
-  if (pure) {
-    return c1 === c2;
-  }
+  const c1: any = React.Children.toArray(_c1);
+  const c2: any = React.Children.toArray(_c2);
   if (c1.length !== c2.length) {
     return false;
   }
   const len = c1.length;
   for (let i = 0; i < len; i++) {
-    if (c1[i].value !== c2[i].value || c1[i].label !== c2[i].label) {
+    if (c1[i].props.value !== c2[i].props.value || c1[i].props.children !== c2[i].props.children) {
       return false;
     }
   }